Resolve Compiler Warnings (#64)

+++ Source/Core/Loader/ERS_ModelLoader/ERS_ModelLoader.cpp
@@ -450,7 +450,7 @@ void ERS_CLASS_ModelLoader::LoadMaterialTextures(std::vector<int>* IDs, std::vec
 
         // Search Texture List For Index Of Same Match, Add -1 If Not Found
         int Index = -1;
-        for (int x = 0; x < TextureList.size(); x++) {
+        for (int x = 0; (long)x < (long)TextureList.size(); x++) {
             if (TextureList[x] == TextureIdentifier) {
                 Index = x;
                 break;
